// src/StreamSource.d.ts
export declare class StreamSource {
    stream: MediaStream;
    source: any;
    isLocal: boolean;
    /**
     *Creates an instance of StreamSource.
     * @param {MediaStream} stream
     * @param {*} source
     * @param {boolean} isLocal
     * @memberof StreamSource
     */
    constructor(stream: MediaStream, source: any, isLocal: boolean);
}